+++ test/bt-bep44.cpp
@@ -165,7 +165,7 @@ int main(int argc, const char** argv)
 
     auto metrics = metrics::Client::noop();
     bool do_doh = true;
-    uint64_t rx_limit = default_udp_mux_rx_limit;
+    uint32_t rx_limit = default_udp_mux_rx_limit;
 
     unique_ptr<MainlineDht> dht(new MainlineDht(ctx.get_executor(), metrics.mainline_dht(), do_doh, rx_limit));
