+++ test/bt-bep5.cpp
@@ -105,7 +105,7 @@ int main(int argc, const char** argv)
     auto metrics_client = metrics::Client::noop();
     auto metrics_dht = metrics_client.mainline_dht();
     bool do_doh = true;
-    uint64_t rx_limit = default_udp_mux_rx_limit;
+    uint32_t rx_limit = default_udp_mux_rx_limit;
 
     DhtNode dht {ctx.get_executor(), metrics_dht.dht_node_ipv4(), do_doh, rx_limit};
